==> R markdown analysis revision/05_plot_20251228.Rmd <==
---
title: "HBV"
output: html_document
date: "2025-12-28"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
library(readxl)
library(openxlsx)

library(patchwork)
```


```{r}
list.files("../plot")
```

```{r}
# read RDS

rds_name <- list.files("../plot") %>% str_subset("RDS") %>% str_remove("\\.RDS")

rds_name
```

```{r}
rds_path <- paste0("../plot/", rds_name, ".RDS")
```

```{r}
for (i in 1:length(rds_name)){
  assign(rds_name[i] %>% str_remove("_[:digit:]+$")　, readRDS(rds_path[i]))
}
```

# read BOJ


```{r}
# 10, 14, 8, 6

p1 <- p1_a1 + p1_a2 + p1_a3 + plot_spacer() +
  p1_b1 +  p1_b2 + p1_b3 + 
  p1_c1 + p1_c2 + plot_spacer() +
  p1_d1 + p1_d2 + p1_d3 +
  plot_layout(design = "
            AABBCCCDDD
            AABBCCCDDD
            AABBCCCDDD
            AABBCCCDDD
            AABBCCCDDD
            AABBCCCDDD
            AABBCCCDDD
            AABBCCCDDD
            AABBCCCDDD
            AABBCCCDDD
            AABBCCCDDD
            AABBCCCDDD
            AABBCCCDDD
            AABBCCCDDD
            AABBCCCDDD
            EEEEFFFGGG
            EEEEFFFGGG
            EEEEFFFGGG
            EEEEFFFGGG
            EEEEFFFGGG
            EEEEFFFGGG
            EEEEFFFGGG
            EEEEFFFGGG
            EEEEFFFGGG
            EEEEFFFGGG
            EEEEFFFGGG
            EEEEFFFGGG
            EEEEFFFGGG
            EEEEFFFGGG
            EEEEFFFGGG
            EEEEFFFGGG
            EEEEFFFGGG
            EEEEFFFGGG
            EEEEFFFGGG
            EEEEFFFGGG
            EEEEFFFGGG
            HHHHIIIJJJ
            HHHHIIIJJJ
            HHHHIIIJJJ
            HHHHIIIJJJ
            HHHHIIIJJJ
            HHHHIIIJJJ
            HHHHIIIJJJ
            HHHHIIIJJJ
            HHHHIIIJJJ
            HHHHIIIJJJ
            HHHHIIIJJJ
            HHHHIIIJJJ
            KKKKLLLMMM
            KKKKLLLMMM
            KKKKLLLMMM
            KKKKLLLMMM
            KKKKLLLMMM
            KKKKLLLMMM
            KKKKLLLMMM
            KKKKLLLMMM
            KKKKLLLMMM")

ggsave("../figure/Fig1.pdf", p1, width = 174, height = 234, unit = "mm")
```
